+++ .env.example
@@ -29,3 +29,12 @@ AGE_GRAPH_NAME=accounting_graph
 # [애플리케이션 전용 로깅 및 추가 설정]
 # LOG_LEVEL: 애플리케이션에서 출력하는 터미널 메시지(디버깅 로그)의 상세도(INFO, DEBUG, ERROR)를 환경 변수만으로 손쉽게 조절하기 위해 마련된 플레이스홀더입니다.
 # LOG_LEVEL=INFO
+
+# [Reranking Configuration]
+# USE_RERANKER: Cross-Encoder 리랭킹 모델 활성화 여부입니다.
+#   - false(기본값): 1차 검색 결과를 그대로 사용하며 리랭킹 모델을 호출하지 않습니다.
+#   - true: 검색 결과를 리랭킹 모델로 재정렬한 후 임계값 필터링을 수행합니다.
+# RERANK_THRESHOLD: USE_RERANKER=true일 때 적용되는 리랭킹 점수 임계값입니다 (0.0 ~ 1.0)
+#   최고 점수가 이 값 미만이면 재검색(CRAG 루프)을 트리거합니다.
+USE_RERANKER=false
+RERANK_THRESHOLD=0.5
