+++ src/utils/config.py
@@ -3,7 +3,10 @@ from datetime import timezone, timedelta
 # 파이프라인 전역 설정값
 MAX_REWRITE_COUNT: int = 3          # FUNC-004: 질의 재작성 최대 반복 횟수
 TOP_K_RETRIEVAL: int = 10           # FUNC-005: 1차 검색 반환 청크 수
-RERANK_THRESHOLD: float = 0.5       # FUNC-006: 재정렬 후 필터링 임계값
+
+# Reranking Configuration
+USE_RERANKER: bool = False          # 리랭킹 모델 활성화 여부
+RERANK_THRESHOLD: float = 0.5       # FUNC-006: 재정렬 후 필터링 임계값 (기본값: 중간 신뢰도)
 VECTOR_COLLECTION_NAME: str = "rag_for_accounting"  # FUNC-003: pgvector 컬렉션명
 OPENAI_MODEL: str = "gpt-5.4-mini"   # FUNC-007, 008, 009: LLM 모델 식별자
